Bump RocksDB version from 4.8 to 4.13.5

+++ src/suffixarray-phrasetable/native/suffixarray/UpdateBatch.cpp
@@ -22,7 +22,7 @@ void UpdateBatch::Reset(const vector<seqid_t> &_streams) {
 }
 
 bool UpdateBatch::SetStreamIfValid(stream_t stream, seqid_t sentence) {
-    if (streams.size() <= stream)
+    if ((int) streams.size() <= stream)
         streams.resize(((size_t) stream) + 1, -1);
 
     if (streams[stream] < sentence) {
